AI-chat/backend/requirements.txt
Flask==2.1.2
flask-cors==3.0.10
google-genai==1.0.0
Werkzeug==2.0.3
gunicorn==20.1.0